1, Inconsistence process of binlog producing and binlog consuming when only kHeaderSize left in current block\n2, fix append blank put more than one kHeaderSize problem\n3, update slash

+++ src/node/zp_data_client_conn.cc
@@ -28,7 +28,8 @@ int ZPDataClientConn::DealMessage() {
 // Msg is  [ length (int32) | pb_msg (length bytes) ]
 int ZPDataClientConn::DealMessageInternal() {
   if (!zp_data_server->Availible()) {
-    LOG(WARNING) << "Receive Client command " << request_.type() << " from (" << ip_port() << "), but the server is not availible yet";
+    LOG(WARNING) << "Receive Client command " << static_cast<int>(request_.type())
+      << " from (" << ip_port() << "), but the server is not availible yet";
     response_.set_type(request_.type());
     response_.set_code(client::StatusCode::kError);
     response_.set_msg("server is not availible yet");
